标准化 int 为 int32_t

+++ fibjs/include/Redis.h
@@ -77,7 +77,7 @@ public:
     virtual result_t close();
 
 public:
-    result_t connect(const char *host, int port, AsyncEvent *ac);
+    result_t connect(const char *host, int32_t port, AsyncEvent *ac);
     result_t _command(std::string &req, Variant &retVal, AsyncEvent *ac);
     ASYNC_MEMBERVALUE2(Redis, _command, std::string, Variant);
 
@@ -134,7 +134,7 @@ public:
 #ifdef _WIN32
             sprintf(numStr, "%I64d", v);
 #else
-            sprintf(numStr, "%lld", (long long)v);
+            sprintf(numStr, "%lld", (int64_t)v);
 #endif
             return add(numStr);
         }
